Refactor binary position (fixes #10)
The binary reader object is now simply a file like object.

There is no longer any concept of a position, and a single binary reader object
is used to represent the file like object. It has all the properties of the old
binaryPosition, and has the begin, end, pos functionality of the old
BinaryReader.

elfParser no longer builds its own file object: This needlessly forces us to
template the class. A binaryReader object is versatile enough.

This change has force a minor refactoring (for most files nothing more than
changing the name of a type)  across the libLink library