Add the ability to pull string to the binary reader
Added Pull() and >> to the BinaryReader.

They calcualte the length of the string, read it into the pointer and
advance the the offset of the reader.

A separate template is designed to take any piece of POD, calculate its
size and reads into it, again advancing the offset. Tests / bug fixes
for this will follow in a seprate commit